requirements.txt
aiogram==3.13.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
apscheduler==3.10.4
pytz==2024.2
python-dotenv==1.0.1